add Policy

+++ app/Http/Controllers/WorkerController.php
@@ -6,11 +6,12 @@ use App\Http\Requests\Worker\FilterRequest;
 use App\Http\Requests\Worker\StoreRequest;
 use App\Http\Requests\Worker\UpdateRequest;
 use App\Models\Worker;
-use Doctrine\Inflector\Rules\Word;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class WorkerController extends Controller
 {
+  use AuthorizesRequests;
   public function index(FilterRequest $request)
   {
     // $workers = Worker::all();
@@ -53,6 +54,7 @@ class WorkerController extends Controller
   }
   public function create()
   {
+    Gate::authorize('create', Worker::class);
     return view('worker.create');
     // $worker = [
     //   'name' => 'Artem',
@@ -67,6 +69,8 @@ class WorkerController extends Controller
   }
   public function store(StoreRequest $request)
   {
+    // $this->authorize('create', Worker::class);
+    Gate::authorize('create', Worker::class);
     $data = $request->validated();
     $data['is_married'] = isset($data['is_married']);
     Worker::create($data);
@@ -74,10 +78,12 @@ class WorkerController extends Controller
   }
   public function edit(Worker $worker)
   {
+    Gate::authorize('update', $worker);
     return view('worker.edit', compact('worker'));
   }
   public function update(UpdateRequest $request, Worker $worker)
   {
+    Gate::authorize('update', $worker);
     $data = $request->validated();
     $data['is_married'] = isset($data['is_married']);
     $worker->update($data);
@@ -96,6 +102,7 @@ class WorkerController extends Controller
   }
   public function destroy(Worker $worker)
   {
+    Gate::authorize('delete', $worker);
     $worker->delete();
     return redirect()->route('workers.index');
   }
